moved common functions to marketplace services (#10319)
