+++ Tests/Marketplace/copy_and_upload_packs.py
@@ -11,9 +11,8 @@ from google.cloud.storage import Blob, Bucket
 
 from Tests.scripts.utils.log_util import install_logging
 from Tests.Marketplace.marketplace_services import init_storage_client, Pack, PackStatus, GCPConfig, PACKS_FULL_PATH, \
-    IGNORED_FILES, PACKS_FOLDER, PACKS_RESULTS_FILE
-from Tests.Marketplace.upload_packs import extract_packs_artifacts, print_packs_summary, load_json, \
-    get_packs_summary
+    IGNORED_FILES, PACKS_FOLDER, PACKS_RESULTS_FILE, load_json
+from Tests.Marketplace.upload_packs import extract_packs_artifacts, print_packs_summary, get_packs_summary
 
 LATEST_ZIP_REGEX = re.compile(fr'^{GCPConfig.GCS_PUBLIC_URL}/[\w./-]+/content/packs/([A-Za-z0-9-_.]+/\d+\.\d+\.\d+/'
                               r'[A-Za-z0-9-_.]+\.zip$)')
